*This* should stop MSVC from bitching.

+++ source/core/StarStringView.cpp
@@ -363,11 +363,11 @@ StringView StringView::substr(size_t position, size_t n) const {
 
   for (size_t i = 0; i != n; ++i) {
     if (it == itEnd)
-      return StringView(start, it.base() - start);
+      return StringView(start, (size_t)(it.base() - start));
     ++it;
   }
 
-  return StringView(start, it.base() - start);
+  return StringView(start, (size_t)(it.base() - start));
 }
 
 int StringView::compare(size_t selfOffset, size_t selfLen, StringView other,
